# api/routers/events.py
"""Events router."""

from datetime import datetime

from fastapi import APIRouter, Depends, HTTPException, Query, status
from pydantic import BaseModel
from sqlalchemy.orm import Session

from api.database import get_db
from api.dependencies import get_current_admin_user, verify_org_member
from api.models import (
    Assignment,
    Event,
    EventTeam,
    Organization,
    Person,
    Team,
)
from api.schemas.common import PaginationParams, get_pagination_params
from api.schemas.event import EventCreate, EventList, EventResponse, EventUpdate
from api.utils.event_helpers import (
    count_people_with_role,
    get_assigned_person_ids,
    get_blocked_assigned_people,
    get_event_required_roles,
    is_person_blocked_on_date,
    person_has_matching_role,
    validate_time_range,
)
from api.utils.response_messages import error_response, success_response, validation_warning

router = APIRouter(prefix="/events", tags=["events"])


class AvailablePerson(BaseModel):
    """Person available for an event."""

    id: str
    name: str
    email: str | None
    roles: list[str]
    is_assigned: bool
    is_blocked: bool = False  # True if person has blocked this date


class AssignmentRequest(BaseModel):
    """Request to assign/unassign a person."""

    person_id: str
    action: str  # "assign" or "unassign"
    role: str | None = None  # Event-specific role (e.g., "usher", "greeter", "sound_tech")


@router.post("/", response_model=EventResponse, status_code=status.HTTP_201_CREATED)
def create_event(
    event_data: EventCreate,
    current_admin: Person = Depends(get_current_admin_user),
    db: Session = Depends(get_db),
):
    """Create a new event (admin only)."""
    # Verify admin belongs to the organization
    verify_org_member(current_admin, event_data.org_id)

    # Verify organization exists
    org = db.query(Organization).filter(Organization.id == event_data.org_id).first()
    if not org:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=f"Organization '{event_data.org_id}' not found",
        )

    # Check if event already exists
    existing = db.query(Event).filter(Event.id == event_data.id).first()
    if existing:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail=f"Event with ID '{event_data.id}' already exists",
        )

    # Validate times
    is_valid, error_message = validate_time_range(event_data.start_time, event_data.end_time)
    if not is_valid:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=error_message,
        )

    # Create event
    event = Event(
        id=event_data.id,
        org_id=event_data.org_id,
        type=event_data.type,
        start_time=event_data.start_time,
        end_time=event_data.end_time,
        resource_id=event_data.resource_id,
        extra_data=event_data.extra_data or {},
    )
    db.add(event)
    db.flush()

    # Add event teams
    if event_data.team_ids:
        for team_id in event_data.team_ids:
            # Verify team exists
            team = db.query(Team).filter(Team.id == team_id).first()
            if not team:
                db.rollback()
                raise HTTPException(
                    status_code=status.HTTP_404_NOT_FOUND,
                    detail=f"Team '{team_id}' not found",
                )
            event_team = EventTeam(event_id=event.id, team_id=team_id)
            db.add(event_team)

    db.commit()
    db.refresh(event)
    return event


@router.get("/", response_model=EventList)
def list_events(
    org_id: str | None = Query(None, description="Filter by organization ID"),
    event_type: str | None = Query(None, description="Filter by event type"),
    start_after: datetime
    | None = Query(None, description="Filter events starting after this time"),
    start_before: datetime
    | None = Query(None, description="Filter events starting before this time"),
    pagination: PaginationParams = Depends(get_pagination_params),
    db: Session = Depends(get_db),
):
    """List events with optional filters."""
    query = db.query(Event)

    if org_id:
        query = query.filter(Event.org_id == org_id)
    if event_type:
        query = query.filter(Event.type == event_type)
    if start_after:
        query = query.filter(Event.start_time >= start_after)
    if start_before:
        query = query.filter(Event.start_time <= start_before)

    query = query.order_by(Event.start_time)
    events = query.offset(pagination.offset).limit(pagination.limit).all()
    total = query.count()
    return {
        "items": events,
        "total": total,
        "limit": pagination.limit,
        "offset": pagination.offset,
    }


@router.get("/{event_id}", response_model=EventResponse)
def get_event(event_id: str, db: Session = Depends(get_db)):
    """Get event by ID."""
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Event '{event_id}' not found"
        )
    return event


@router.put("/{event_id}", response_model=EventResponse)
def update_event(
    event_id: str,
    event_data: EventUpdate,
    current_admin: Person = Depends(get_current_admin_user),
    db: Session = Depends(get_db),
):
    """Update event (admin only)."""
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Event '{event_id}' not found"
        )

    # Verify admin belongs to the same organization as the event
    verify_org_member(current_admin, event.org_id)

    # Update fields
    if event_data.type is not None:
        event.type = event_data.type
    if event_data.start_time is not None:
        event.start_time = event_data.start_time
    if event_data.end_time is not None:
        event.end_time = event_data.end_time
    if event_data.resource_id is not None:
        event.resource_id = event_data.resource_id
    if event_data.extra_data is not None:
        event.extra_data = event_data.extra_data

    # Validate times
    is_valid, error_message = validate_time_range(event.start_time, event.end_time)
    if not is_valid:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=error_message,
        )

    db.commit()
    db.refresh(event)
    return event


@router.delete("/{event_id}", status_code=status.HTTP_204_NO_CONTENT)
def delete_event(
    event_id: str,
    current_admin: Person = Depends(get_current_admin_user),
    db: Session = Depends(get_db),
):
    """Delete event (admin only)."""
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Event '{event_id}' not found"
        )

    # Verify admin belongs to the same organization as the event
    verify_org_member(current_admin, event.org_id)

    db.delete(event)
    db.commit()
    return None


@router.get("/{event_id}/available-people", response_model=list[AvailablePerson])
def get_available_people(event_id: str, db: Session = Depends(get_db)) -> list[AvailablePerson]:
    """
    Get people available for this event based on roles.

    Returns list of people who have matching roles, with flags indicating
    if they're already assigned or have blocked the event date.
    """
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Event '{event_id}' not found"
        )

    # Get required roles from event
    required_roles = get_event_required_roles(event)

    # Get all people in the organization
    people = db.query(Person).filter(Person.org_id == event.org_id).all()

    # Get current assignments for this event
    assigned_person_ids = get_assigned_person_ids(db, event_id)

    # Get event date (just the date part, not time)
    event_date = event.start_time.date()

    # Filter people by matching roles
    available = []
    for person in people:
        person_roles = person.roles or []

        # Check if person has any of the required roles
        if person_has_matching_role(person_roles, required_roles):
            # Check if person has blocked this date
            is_blocked = is_person_blocked_on_date(db, person.id, event_date)

            available.append(
                AvailablePerson(
                    id=person.id,
                    name=person.name,
                    email=person.email,
                    roles=person_roles,
                    is_assigned=person.id in assigned_person_ids,
                    is_blocked=is_blocked,
                )
            )

    return available


@router.get("/{event_id}/validation")
def validate_event(event_id: str, db: Session = Depends(get_db)) -> dict:
    """
    Validate if event has proper configuration and enough people.

    Checks:
    1. Event has role requirements configured
    2. Enough people available for each role
    3. No assigned people are blocked on the event date

    Returns:
        Dictionary with is_valid flag and list of validation warnings
    """
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Event '{event_id}' not found"
        )

    warnings = []
    is_valid = True

    # Check if event has role_counts
    role_counts = (event.extra_data or {}).get("role_counts", {})

    if not role_counts:
        warnings.append(validation_warning("missing_config", "events.validation.no_roles"))
        is_valid = False
    else:
        # Check each role has enough people
        people = db.query(Person).filter(Person.org_id == event.org_id).all()

        for role, needed_count in role_counts.items():
            # Count people with this role
            available_count = count_people_with_role(people, role)

            if available_count < needed_count:
                warnings.append(
                    validation_warning(
                        "insufficient_people",
                        "events.validation.need_more_people",
                        needed=needed_count,
                        role=role,
                        available=available_count,
                    )
                )
                is_valid = False

    # Check if any assigned people are blocked on this event date
    event_date = event.start_time.date()
    blocked_people = get_blocked_assigned_people(db, event_id, event_date)

    if blocked_people:
        warnings.append(
            validation_warning(
                "blocked_assignments",
                "events.validation.blocked_people_assigned",
                people=", ".join(blocked_people),
            )
        )
        is_valid = False

    return {"event_id": event_id, "is_valid": is_valid, "warnings": warnings}


@router.post("/{event_id}/assignments")
def manage_assignment(
    event_id: str,
    request: AssignmentRequest,
    current_admin: Person = Depends(get_current_admin_user),
    db: Session = Depends(get_db),
):
    """Assign or unassign a person to/from an event (admin only)."""
    event = db.query(Event).filter(Event.id == event_id).first()
    if not event:
        raise error_response("events.errors.event_not_found", status_code=status.HTTP_404_NOT_FOUND)

    # Verify admin belongs to the same organization as the event
    verify_org_member(current_admin, event.org_id)

    person = db.query(Person).filter(Person.id == request.person_id).first()
    if not person:
        raise error_response(
            "events.errors.person_not_found", status_code=status.HTTP_404_NOT_FOUND
        )

    # Verify person belongs to the same organization
    verify_org_member(current_admin, person.org_id)

    if request.action == "assign":
        # Check if already assigned
        existing = (
            db.query(Assignment)
            .filter(Assignment.event_id == event_id, Assignment.person_id == request.person_id)
            .first()
        )

        if existing:
            raise error_response(
                "events.assign.already_assigned",
                status_code=status.HTTP_400_BAD_REQUEST,
                person=person.name,
            )

        # Create new assignment (solution_id is None for manual assignments)
        assignment = Assignment(
            event_id=event_id,
            person_id=request.person_id,
            role=request.role,  # Event-specific role
            solution_id=None,
        )
        db.add(assignment)
        db.commit()
        db.refresh(assignment)  # Refresh to get assignment ID

        return success_response(
            "events.assign.success",
            {"assignment_id": assignment.id, "role": request.role},
            person=person.name,
        )

    elif request.action == "unassign":
        # Find and delete assignment
        assignment = (
            db.query(Assignment)
            .filter(Assignment.event_id == event_id, Assignment.person_id == request.person_id)
            .first()
        )

        if not assignment:
            raise error_response(
                "events.assign.not_assigned",
                status_code=status.HTTP_404_NOT_FOUND,
                person=person.name,
            )

        db.delete(assignment)
        db.commit()
        return success_response("events.assign.unassigned", person=person.name)

    else:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=f"Invalid action '{request.action}'. Must be 'assign' or 'unassign'",
        )


@router.get("/assignments/all")
def get_all_assignments(
    org_id: str = Query(..., description="Organization ID"), db: Session = Depends(get_db)
):
    """Get all assignments for an organization (both from solutions and manual)."""
    # Verify organization exists
    org = db.query(Organization).filter(Organization.id == org_id).first()
    if not org:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND, detail=f"Organization '{org_id}' not found"
        )

    # Get all assignments joined with events to filter by organization
    assignments = db.query(Assignment).join(Event).filter(Event.org_id == org_id).all()

    result = []
    for assignment in assignments:
        event = db.query(Event).filter(Event.id == assignment.event_id).first()
        person = db.query(Person).filter(Person.id == assignment.person_id).first()

        result.append(
            {
                "assignment_id": assignment.id,
                "event_id": assignment.event_id,
                "event_type": event.type if event else None,
                "event_start": event.start_time if event else None,
                "event_end": event.end_time if event else None,
                "person_id": assignment.person_id,
                "person_name": person.name if person else None,
                "role": assignment.role,  # Event-specific role
                "solution_id": assignment.solution_id,
                "is_manual": assignment.solution_id is None,
            }
        )

    return {"assignments": result, "total": len(result)}