# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AssignmentRequest": {
        "description": "Request to assign/unassign a person.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "person_id": {
            "title": "Person Id",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "required": [
          "person_id",
          "action"
        ],
        "title": "AssignmentRequest",
        "type": "object"
      },
...
      "EventCreate": {
        "description": "Schema for creating an event.",
        "properties": {
          "end_time": {
            "description": "Event end time",
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "extra_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional data",
            "title": "Extra Data"
          },
          "id": {
            "description": "Unique event ID",
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "description": "Organization ID",
            "title": "Org Id",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resource/venue ID",
            "title": "Resource Id"
          },
          "start_time": {
            "description": "Event start time",
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "team_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "List of team IDs",
            "title": "Team Ids"
          },
          "type": {
            "description": "Event type (match, shift, meeting)",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "start_time",
          "end_time",
          "id",
          "org_id"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventResponse": {
        "description": "Schema for event response.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "end_time": {
            "description": "Event end time",
            "format": "date-time",
            "title": "End Time",
            "type": "string"
          },
          "extra_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional data",
            "title": "Extra Data"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resource/venue ID",
            "title": "Resource Id"
          },
          "start_time": {
            "description": "Event start time",
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "type": {
            "description": "Event type (match, shift, meeting)",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "type",
          "start_time",
          "end_time",
          "id",
          "org_id",
          "created_at",
          "updated_at"
        ],
        "title": "EventResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_EventResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ListResponse[EventResponse]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered volunteer scheduling and sign-up management",
    "title": "SignUpFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events/": {
      "get": {
        "description": "List events with optional filters.",
        "operationId": "listEvents",
        "parameters": [
          {
            "description": "Filter by organization ID",
            "in": "query",
            "name": "org_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization ID",
              "title": "Org Id"
            }
          },
          {
            "description": "Filter by event type",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by event type",
              "title": "Event Type"
            }
          },
          {
            "description": "Filter events starting after this time",
            "in": "query",
            "name": "start_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter events starting after this time",
              "title": "Start After"
            }
          },
          {
            "description": "Filter events starting before this time",
            "in": "query",
            "name": "start_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter events starting before this time",
              "title": "Start Before"
            }
          },
          {
            "description": "Page size, max 200",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size, max 200",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of rows to skip",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of rows to skip",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_EventResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Create a new event (admin only).",
        "operationId": "createEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/assignments/all": {
      "get": {
        "description": "Get all assignments for an organization (both from solutions and manual).",
        "operationId": "getAllAssignments",
        "parameters": [
          {
            "description": "Organization ID",
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "description": "Organization ID",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get All Assignments",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "delete": {
        "description": "Delete event (admin only).",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "get": {
        "description": "Get event by ID.",
        "operationId": "getEvent",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event",
        "tags": [
          "events"
        ]
      },
      "put": {
        "description": "Update event (admin only).",
        "operationId": "updateEvent",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/assignments": {
      "post": {
        "description": "Assign or unassign a person to/from an event (admin only).",
        "operationId": "manageAssignment",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Manage Assignment",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/available-people": {
      "get": {
        "description": "Get people available for this event based on roles.\n\nReturns list of people who have matching roles, with flags indicating\nif they're already assigned or have blocked the event date.",
        "operationId": "getAvailablePeople",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AvailablePerson"
                  },
                  "title": "Response Get Available People Api V1 Events  Event Id  Available People Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Available People",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/validation": {
      "get": {
        "description": "Validate if event has proper configuration and enough people.\n\nChecks:\n1. Event has role requirements configured\n2. Enough people available for each role\n3. No assigned people are blocked on the event date\n\nReturns:\n    Dictionary with is_valid flag and list of validation warnings",
        "operationId": "validateEvent",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Event Api V1 Events  Event Id  Validation Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Event",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}